+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `Drop` but instead only asserts that every field implements `ZeroizeOnDrop`.
 
 ### Fixed
-- Stop depending on unstable APIs of `zeroize` to implement `ZeroizeOnDrop`.
+- Stop depending on unstable APIs for `Eq` for `ZeroizeOnDrop`.
 
 ## [1.4.0] - 2025-05-01